Optimal code?
Removing/changing a nummber of comments, updating some formatting, changing various small things. This may or may not be the optimal version of the code, but it's closer to it.
Since our robot didn't really work, we removed a lot of things from the final code. This adds them back in some way. It probably could be better but it doesn't matter that much.